Add in spin wheel's css, and private vars. Fix test code.
This adapted to be fully based on spin wheel code now.

For adding private member variables inside polymer class, we do it by
adding inside ready() function inside beforeRegister() function.

Actually we can also do it by adding them inside this.properties, then
prefix it with _ and mark readOnly to true. The former should be really
pure javascript and not to mingle with polymer in anyway. Thus we chose
to do that.

Known issue

* css won't be rendered yet as the syntax is based on scss. We will add
in another tool to convert scss to css and separate style into another
file to fix it.