Local style to html element
Placed <style></style> inside <template></template> to get the effect of
style to local DOM as stated here

"Note: Prior to Polymer 1.1, the recommendation was to place <style>
tags inside the <dom-module> for an element (but outside the
<template>). This is still supported, but is no longer recommended."

Note also that we need to remove pb-spinwheel element name from css too.

With this change, we no longer need to adding another tool to convert
scss to css as it will still be local DOM, no need for better syntax
support for convenient at this point.

Ref:
https://www.polymer-project.org/1.0/docs/devguide/styling