Fix code to reflect to have Playbasis built first
Added kErrorCode with initially one error in place for having non-built
Playbasis-environment just yet. If that happened, then it will fire
error event with that error code.

Event listeners for success and error event are changed to be added
outside of WebComponentsReady event as it can happen before that event
is fired; as in our case of non-built-Playbasis-environment.

Error information is created via Error class with customized .code
property to set to a proper error code as seen in kErrorCode of
component property.