Refactor code. Fire error event as necessary for branch that needs it.
This changed to use arrow function as necessary but avoid using on
Polymer's properties.

Firing error event when necessary for locations that need it. This
does't apply for result of function that return promise, as it's still
in the mid of communication, this will apply only for the end-point of
function calling.