Fix promise error handler. Add setting player id from HTML's attribute.
For promise error handler, use

funct().then((r) => {}, (e) => {});

pattern instead of

funct().then((r) => {}).error((e) => {});

in which the latter case might be missed to called if multiple different
Promise class is used i.e. native Promise class, and bluebird Promise
class. There's a sligtly differences that is best to avoid by using the
former pattern for better.

Player id now can be set via HTML element's attribute as follows

<pb-spinwheel player-id="your-player-id"></pb-spinwheel>

it's set to null as default, and in run time will be checked if it has a
proper value or not (not null or empty), if so it will reject as part of
Promise object return of function call.

Also more comments are added in index.html for more information about
big steps in using the component.